Added condition to volume exporter for cleanup.

+++ docker/captain_volume_exporter/latest/files/scripts/run.sh
@@ -30,8 +30,10 @@ run_script () {
       tar cvf "$tar_file_with_path" "$VOLUME"
 
       if [[ ! -z "$AWS_S3_ACCESS_KEY_ID" ]]; then
-        echo "removing old local backup file $old_tar_file_with_path..."
-        rm -f "$old_tar_file_with_path"
+        if [[ ! -z "$old_tar_file_with_path" ]]; then
+          echo "removing old local backup file $old_tar_file_with_path..."
+          rm -f "$old_tar_file_with_path"
+        fi
         echo "sending $tar_file_with_path to AWS S3 bucket $AWS_S3_BUCKET_NAME on path $AWS_S3_PATH..."
         local aws_file="$AWS_S3_PATH""$tar_file"
         upload_to_aws_s3 "$AWS_S3_ACCESS_KEY_ID" "$AWS_S3_SECRET_ACCESS_KEY" "$AWS_S3_BUCKET_NAME" "$AWS_S3_BUCKET_REGION" "$aws_file" "$tar_file_with_path"
